packages/dbgear/pyproject.toml
[tool.poetry]
name = "dbgear"
version = "0.14.3"
description = "Database management tools for initial data management"
authors = ["tamuto <tamuto@infodb.jp>"]
readme = "README.md"
homepage = "https://github.com/tamuto/dbgear"
repository = "https://github.com/tamuto/dbgear"

packages = [
    { include = "dbgear", from = "." }
]

[tool.poetry.scripts]
dbgear = "dbgear.main:execute"

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0"
sqlalchemy = "^2.0.38"
pymysql = "^1.1.1"
python-ulid = "^2.5.0"
cryptography = "^42.0.7"
pydantic = "^2.0.0"
jinja2 = "^3.1.6"

[tool.poetry.group.dev.dependencies]
flake8 = "^3.9.2"
taskipy = "^1.12.0"

[tool.poetry.group.plugin.dependencies]
echo-plugin = {path = "../../etc/plugins/echo"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
test = "poetry run python -m unittest discover tests/ -v"
test-fast = "poetry run python -m unittest discover tests/ -k 'not slow'"
lint = "poetry run flake8 dbgear/ tests/"
clean = "rm -rf dist/ build/ *.egg-info/"
cli = "poetry run dbgear"
